requirements.txt
aiohttp
APScheduler
async-timeout
chardet
discord.py
emoji
multidict
pytz
six
tzlocal
websockets
bs4
lxml
requests